# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "Documentation de l'API pour la gestion des articles, utilisateurs, clients et rôles."
  },
  "servers": [
  {
    "url": "http://127.0.0.1:8000/api/v1",
    "description": "Serveur de production"
  }
],
  "paths": {
...
    "/clients": {
      "get": {
        "summary": "Lister les clients",
        "operationId": "indexbis",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des clients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "clients": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Client"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Aucun client trouvé"
          }
        }
      },
      "post": {
        "summary": "Créer un client",
        "operationId": "createClient",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          },
          "403": {
            "description": "Non autorisé"
          }
        }
      }
    },
...
    "/clients/telephone": {
      "post": {
        "summary": "Obtenir les clients par téléphone",
        "operationId": "getClientsByTelephones",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelephoneRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Clients trouvés par téléphone",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "clients": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Client"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "ClientCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": ["name", "phone", "email"]
      },
...
      "TelephoneRequest": {
        "type": "object",
        "properties": {
          "phone_numbers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["phone_numbers"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}